requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.1.0
pytest==8.3.2
pytest-django==4.5.2

